+++ test/common/__init__.py
@@ -1,2 +1,2 @@
-from .commons import *
+from .common import *
 from .reference_jpeg import entropy_encode, numofbits
