+++ test/common/reference_jpeg.py
