+++ test/test_entropycoder.py
@@ -83,5 +83,3 @@ def test_block_conversion():
 
     verify.simulator = 'iverilog'
     assert bench_entropycoder().verify_convert() == 0
-
-test_entropycoder()
